+++ engine/entities/Enemy.h
@@ -17,8 +17,8 @@ protected:
 
 
 public:
-	Enemy(int id, float speed, float size, int sides, EntityType type, float angle); // used to set in which direction the enemy will move better for coop allows sync
-	Enemy(int id, float speed, float size, int sides, EntityType type); // Better for single player no need to wory about the angle at which the enemy goes
+	Enemy(int id, float speed, float size, float sides, EntityType type, float angle); // used to set in which direction the enemy will move better for coop allows sync
+	Enemy(int id, float speed, float size, float sides, EntityType type); // Better for single player no need to wory about the angle at which the enemy goes
 	void update(float deltaTime);
 	//bool colide(vec2& playerPos);
 	void draw(sf::RenderWindow& window);
